classdojo-debit-system/client/requirements.txt
requests==2.31.0
PyQt5==5.15.9
pyserial==3.5
pyautogui==0.9.54
keyboard==0.13.5
psutil==5.9.6
schedule==1.2.0
